requirements.txt
numpy==1.22.2
pandas==1.4.1
Pillow==9.4.0
Plost==0.2.5
plotly==5.13.1
streamlit==1.20.0
